ActiveModel types should only cast and not validate
