+++ app/controllers/maintenance_tasks/tasks_controller.rb
@@ -33,11 +33,11 @@ module MaintenanceTasks
       redirect_to(task_path(task))
     rescue ActiveRecord::RecordInvalid => error
       redirect_to(task_path(error.record.task_name), alert: error.message)
-    rescue Runner::EnqueuingError => error
-      redirect_to(task_path(error.run.task_name), alert: error.message)
-    rescue ActiveRecord::ValueTooLong, Parameters::TypeError => error
+    rescue ActiveRecord::ValueTooLong => error
       task_name = params.fetch(:id)
       redirect_to(task_path(task_name), alert: error.message)
+    rescue Runner::EnqueuingError => error
+      redirect_to(task_path(error.run.task_name), alert: error.message)
     end
 
     private
