--- app/models/maintenance_tasks/parameters/type_error.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-module MaintenanceTasks
-  module Parameters
-    # Error to raise when parameter ActiveModel::Value::Type objects fail
-    # validation.
-    class TypeError < StandardError; end
-  end
-end
