+++ app/models/maintenance_tasks/task/string_array_type.rb
@@ -1,30 +1,17 @@
 # frozen_string_literal: true
 module MaintenanceTasks
-  module Parameters
+  class Task
     # Type class representing an array of strings. Tasks using the Attributes
     # API for parameter support can use this class to turn input from the UI
     # into an array of strings within their Task.
     class StringArrayType < ActiveModel::Type::Value
-      # Casts string from form input field to an array of strings after
-      # validating that the input is correct.
+      # Casts string from form input field to an array of strings.
       #
       # @param input [String] the value to cast to an array of strings.
       # @return [Array<Integer>] the data cast as an array of strings.
       def cast(input)
         return unless input.present?
-
-        validate_input(input)
         input.split(",")
-      end
-
-      private
-
-      def validate_input(input)
-        unless /\A(\s?\w+\s?(,\s?\w+\s?)*)\z/.match?(input)
-          error_message = "#{self.class} expects alphanumeric, "\
-            "comma-delimited string. Input received: #{input}"
-          raise TypeError, error_message
-        end
       end
     end
   end
