+++ test/models/maintenance_tasks/task/integer_array_type_test.rb
@@ -2,7 +2,7 @@
 require "test_helper"
 
 module MaintenanceTasks
-  module Parameters
+  class Task
     class IntegerArrayTypeTest < ActiveSupport::TestCase
       test "#cast returns nil if input not present" do
         assert_nil IntegerArrayType.new.cast("")
@@ -16,17 +16,8 @@ module MaintenanceTasks
         assert_equal [123, 456, 78], IntegerArrayType.new.cast("123 ,456, 78 ")
       end
 
-      test "#cast raises if input is not valid" do
-        input = "abc"
-        error = assert_raises(TypeError) do
-          IntegerArrayType.new.cast(input)
-        end
-
-        expected_error_message = <<~MSG.squish
-          MaintenanceTasks::Parameters::IntegerArrayType expects a
-          comma-delimited string of integers. Input received: #{input}
-        MSG
-        assert_equal(expected_error_message, error.message)
+      test "#cast handles non-integer input" do
+        assert_equal [0], IntegerArrayType.new.cast("abc")
       end
     end
   end
